+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - qpSWIFT 0.0.2 solver interface
 
+### Changed
+
+- Compute colors automatically in benchmark example
+
 ## [1.8.1] - 2022/03/05
 
 ### Added